+++ HISTORY.md
@@ -2,7 +2,7 @@
 
 ## v0.13.0 3 July 2026
 
-Improve tab/shift-tab cycling and add cursor key navigation.
+- Improve tab/shift-tab cycling and add cursor key navigation.
 
 ## v0.12.2 6 June 2025